added normed version of new exec
