+++ main_work/incs/mini_execution.h
@@ -6,7 +6,7 @@
 /*   By: ljohnson <ljohnson@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/08 08:00:13 by ljohnson          #+#    #+#             */
-/*   Updated: 2022/05/19 14:34:57 by ljohnson         ###   ########lyon.fr   */
+/*   Updated: 2022/05/20 08:05:32 by ljohnson         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,7 +26,19 @@
 		EXECUTION FUNCTIONS PROTOTYPES
 *//////////////////////////////////////////////////////////////////////////////
 
+/*-------------------- exec_built_in.c --------------------*/
+
+int		mini_built_in_part_one(t_master *master, t_cmd *cmd);
+int		mini_built_in_part_two(t_master *master, t_cmd *cmd, int a);
+int		mini_btin_hub(t_master *master, t_cmd *cmd, int pipe_fd[2], int last);
+
 /*-------------------- exec_main.c --------------------*/
+
+int		mini_execve(t_envdata *envdata, t_cmd *cmd);
+int		mini_exec_hub(t_master *master, t_cmd *cmd, int pipe_fd[2], int last);
+int		mini_child_process(t_master *master, t_cmd *cmd, int last);
+int		mini_wait_process(t_master *master);
+int		mini_exec_loop(t_master *master);
 
 /*-------------------- exec_redirection.c --------------------*/
 
@@ -34,24 +46,5 @@ int		mini_set_fd_in(t_cmd *cmd, int a);
 int		mini_set_fd_out(t_cmd *cmd, int a);
 int		mini_redirect(int fd_in, int fd_out);
 int		mini_redirection_hub(t_cmd *cmd);
-
-/*-------------------- /////////////// --------------------*/
-/*-------------------- /////////////// --------------------*/
-/*-------------------- /////////////// --------------------*/
-/*-------------------- /////////////// --------------------*/
-/*-------------------- new_manage_fd.c --------------------*/
-
-int		mini_dup_handler(t_master *master, int pipe_fd[2], int last);
-int		mini_close_child_process(int pipe_fd, int fd_link);
-
-/*-------------------- new_exec.c --------------------*/
-int		mini_built_in_part_one(t_master *master, t_cmd *cmd);
-int		mini_built_in_part_two(t_master *master, t_cmd *cmd, int a);
-int		mini_btin_hub(t_master *master, t_cmd *cmd, int pipe_fd[2], int last);
-
-int		mini_exec_hub(t_master *master, t_cmd *cmd, int pipe_fd[2], int last);
-int		mini_child_process(t_master *master, t_cmd *cmd, int last);
-int		mini_wait_process(t_master *master);
-int		mini_exec_loop(t_master *master);
 
 #endif //MINI_EXECUTION_H
